Revert "Add static FP8 attention support (#1045)"
This reverts commit c5b1c4129021dc2f920673b1011fa7fbad4e6095.
