+++ auto_round/__main__.py
@@ -153,13 +153,6 @@ class BasicArgumentParser(argparse.ArgumentParser):
         basic.add_argument(
             "--enable_torch_compile", action="store_true", help="Enable PyTorch compilation for faster execution. "
         )
-        basic.add_argument(
-            "--static_kv_dtype", default=None, type=str, help="Data type for static quantize key and value. "
-        )
-
-        basic.add_argument(
-            "--static_attention_dtype ", default=None, type=str, help="Data type for static quantize attention. "
-        )
 
         tuning = self.add_argument_group("Tuning Arguments")
         tuning.add_argument(
@@ -606,8 +599,6 @@ def tune(args):
         layer_config=layer_config,
         model_dtype=args.model_dtype,
         momentum=args.momentum,
-        static_kv_dtype=args.static_kv_dtype,
-        static_attention_dtype=args.static_attention_dtype,
     )
 
     model_name = args.model.rstrip("/")
